REF/TST: method-specific files for Series timeseries methods (#32226)
