+++ pandas/tests/series/test_datetime_values.py
@@ -19,7 +19,6 @@ from pandas import (
     PeriodIndex,
     Series,
     TimedeltaIndex,
-    bdate_range,
     date_range,
     period_range,
     timedelta_range,
@@ -621,18 +620,6 @@ class TestSeriesDatetimeValues:
         s.fillna(pd.Timestamp("2018-01-01"), inplace=True)
         result = s.dt.date
         assert result[0] == result[2]
-
-    def test_between(self):
-        s = Series(bdate_range("1/1/2000", periods=20).astype(object))
-        s[::2] = np.nan
-
-        result = s[s.between(s[3], s[17])]
-        expected = s[3:18].dropna()
-        tm.assert_series_equal(result, expected)
-
-        result = s[s.between(s[3], s[17], inclusive=False)]
-        expected = s[5:16].dropna()
-        tm.assert_series_equal(result, expected)
 
     def test_date_tz(self):
         # GH11757
